MSPSDS-491 Fix search results not updating when clearing autocomplete
The autocomplete works by hiding the select and showing an input field
instead. It updates the underlying field when necessary, however it
doesn't do so when the user clear the field and doesn't replace it with
anything else (e.g. selecting 'None'). This was in issue, as the search
results were still using the underyling select - this commit works
around that problem with an extra event listener.