MSPSDS-594 Add a welcome email triggered by required actions reset
The admin might want to configure the newly created user before sending
the welcome email, so triggering a welcome email off of the user-added
event is not an option.

As such, we're re-purposing the existing ability of admins to reset email
verification and other required actions as a welcome email.

If the functionality to trigger password resets with an email sent is
required down the line, then this will have to be decoupled and a
custom email type implemented. This does not impact the
'forgot password' flow, nor the ability to make a user reset their
password on next login.