MSPSDS-802: Add Keycloak resource provider for custom users endpoint
This exposes a new REST API endpoint at /auth/realms/{realm}/admin/users that returns the same response as the standard users admin endpoint, but also includes the group IDs that each user belongs to. The authorisation requirements to access this endpoint are identical to those of the equivalent Keycloak admin REST API users resource.