MSPSDS-896 Revert setting default action and remove it for sms
Setting the required actions to be applied to new accounts by default
is not currently appropriate, as we're planning to integrate with the
office's SSO (okta). Those users should not have to supply the same
details, so having a single set of defaults does not work.

The README setup section for the sms authenticator used to instruct
us to turn providing sms for 2fa as a default required action - this
is now amended to match the rest.