MSPSDS-821 Only run tests on PR builds
The only way that any of the branches we care about change is through a
pull request (modulo admin overriding restrictions).

Therefore, by running the tests on branch changes, we're doubling the
work. This was somewhat mitigated by the `test-if-changed.sh`
script, but the script itself also takes a certain amount of time to
boot up and finish.