MSPSDS-566 Use redis for session data storage
The connection of this change to the msa flow lies in the amount of data
the msa flow wants to store in the session before committing everything
to the db at the end of it. We started running out of space in the 4kb
of the cookie.

Additionally, server-side session stores have
some security advantages over client-side:
https://www.happybearsoftware.com/almost-protect-yourself-from-cookie-session-store