PSD-1176 Make User-TeamUser loading dependency in KC client explicit
This change required some follow up of when we load teams and team
membership from KC. Previously, when loading users, we'd also load
teams and team membership. By following the patter to make these
dependencies explicit, we need to tell `team_users` all users we know,
thus creating a circular dependency. I've opted to add an explicit
load of Team and TeamUser objects in an initializer as a way to break
that cycle - mimicking how we treat Organisations and Users already.