Use webpacker-compatible paths for assets
We aren't using sprockets asset pipeline so the way in which assets were loaded wasn't ideal, as it created a file system location dependency. We want to split shared-web out into a separate repository, so this change will enable it to work regardless of how the assets are loaded via yarn.