Add new poisonous ingredients question
This adds a new question page which lists poisonous ingredients and askes users to say whether their product contains any.

This is only asked if the product is being specified as a frame formulation.

After answering the question, the user is taken to the PH question, skipping the "Does the cosmetic product contain anti-dandruff agents?" question.

The response is also replayed on the check your answers page at the end.