Add pH enum
There are now four possible answers to the pH question, rather than a simple binary yes/no.

Rather than mapping the yes/no to the `applicable` column of the `TriggerQuestion` model, this createas a new enumerated column called `ph` on the `Component` model with 4 possible answers (as well as `null`).