Don't rescue from all errors when parsing CPNP files
Currently if there are _any_ errors when parsing the CPNP files, including syntax errors in our code, we rescue from them and just add an "uknown_error", which displays as:

> Try again or manually enter the production notification data

This means that Sentry isn’t notified of the exception.  By removing this, we'll be alerted to these errors earlier, and can then either fix them or add more specific errors that will make more sense to users.