Remove contact person email verification (#1442)
Removes the need for the e-mail address for the "contact person" to be verified by clicking on a link received via e-mail. This is now deemed unnecessary, and has proved to be an annoying stumbling block in the account creation and set up process (as it may mean a user having to wait for someone else to receive an e-mail).