Don't render error pages for request specs
This changes the behaviour for tests so that error pages are no longer rendered, except for system tests.

This has the advantage that for request specs, if you've got a syntax error or typo in your code, you can see the exception within the console, rather than getting an error based on the application rendering a 500 error page.

The downside is that if we want to test errors (eg 404 not found or 403 access denied) we'd either have to assert that an exception was raised, or change this setting just for those tests so that we can assert based on the http response code or error page content.