Eliminate e-mail address collisions in automated tests
Currently we’re using a shared Keycloak instance for our automated tests, with no regular purging of the database, so it is filling up with test user accounts. This is resulting in random test failures when the e-mail address is already taken.

By changing from a 1 to 100 random number to a number based on the current time (in milliseconds) we should eliminate the chance of a e-mail address collision.