Default to adding a validation on belongs_to relationships
This adds the `belongs_to_required_by_default = true` config option, which is the default for new Rails applications. This adds an automatic validation to all `belongs_to` relationships which presents the foreign key from being null.

By adding this, we can remove the explicit `validates :notification, presence: true` on the Component model.