Adding feature specs for notifications (#1517)
This adds rspec feature specs to test the end-to-end journey of notifying cosmetic products via either uploading a ZIP file or "manually" by answering questions, in both pre and post Brexit scenarios, and with different variants of notification.

As ZIP files and product images are processed by jobs in queue, to make these flows more easily testable, the job queuing process within the test environment is changed from :async (using a background thread) to :inline. As a consequence of this, some of the other background tasks (such as scanning for viruses or sending e-mails) are stubbed to avoid throwing exceptions.