Prevent dependabot suggesting changes to Gemfile
This change will ask Dependabot to only suggest changes to the `Gemfile.lock`, _not_ the `Gemfile` – [see documentation](https://dependabot.com/docs/config-file/#version_requirement_updates).

The advantage of this is it allows us to, for instance, specify that we want the latest `6.*` release of the `elasticsearch` gem, but not a `7.*` release, as that would require also updating the elasticsearch service itself.