Use pessimistic operator in Gemfile
This updates the `Gemfile` to use the [pessimistic operator](https://thoughtbot.com/blog/rubys-pessimistic-operator) (`~>`) which specifies a minimum version of the dependency, but also allows for minor and patch updates.

For example,

`gem "rails", "~> 5.2.3"`

specifies that the `rails` gem must be on _at least_ version `5.2.3`, but must be below `6.*`.

This is based on the assumption that gems follow [semver](https://semver.org) and so major version updates contain breaking backwards-incompatible changes.