Use error summary helper (#1615)
This updates the error summaries on all of the views to use a new `error_summary_for(model)` helper, which in turn uses a new `govukErrorSummary()` helper within the shared [`govuk-design-system-rails`](https://github.com/uKGovernmentBEIS/govuk-design-system-rails) gem.

The main difference is that this new helper follows the recommendations of the GOV.UK Design System in [linking from the error summary to each answer](https://design-system.service.gov.uk/components/error-summary/#linking-from-the-error-summary-to-each-answer). This means that when you click the link within the error summary, the relevant field is focused.