Specify sort order when looping through components and nano elements (#1622)
When uploading a zip file, the application needs to loop you through any components and/or nanomaterials which need additional information. This change makes sure that these are looped through in the database’s `id` order.  When this isn't specified, the order can be unpredictable (based upon Postgres internals), which was causing issues when writing tests.