Link to Submit deployed reviewap in Deploy Status
When a review app has been deployed successfully, the link to the
deployed reviewapp will be shown in the Github Deploy status details
at the Pull Request.

The review app produces 2 website urls:
- Cosmetics Search.
- Cosmetics Submit.

As the Deploy status can only be linked to one URL, I choose the Submit
URL being the one that has more workload.