Set individual Github environments for reviewapp deploys (#1657)
* Set individual environments for reviewapp deploys

All the reviewapps are being deployed under "int" environment on Github
deployments, matching the BEIS PaaS "int" Space that hosts all the
review apps.

This approach has a critical downside though: New PRs are considered as
the latest version of the "int" environment, and Github Deployment tags
the previously deployed PRs as "innactive", removing he deployment
information and environment URL.
As a consequence of this: We cannot have multiple PRs opened in parallel
showing Deployment & Environment information.

In order to avoid this, we are creating individual environments for each
PullRequest number. So multiple review app environments will be
cohabiting in Github Deployment domain without stepping over each other.

* Give explicit flag for deploying transient envs

Instead of deciding if a Github deployment is transient or not based on
the name of the environment, we extract that decision to the workflow
calling the function.

The deployment creation function will receive an optional flag that will
indicate if the environment is transient or not.

* Extract log URLs from Github Deployment script

The urls to track the deployment process were hardcoded in the
deployment functions, where the url was chosen depending on the
environment name.

This creates coupling between the deployment functions and the
workflows. Extracting them we achieve a less coupled codebase.