Fix missing urls in staging and production deploys
When the URLS where extracted from the Bash deploy functions to the
workflows, I forgot to pass them as parameter to the functions when
calling them from Staging/Production deploy workflow.

As a result, the environment URL is showing up as the log URL, while
the log URL to track the deploy process is not being displayed at all.

This change adds the missing argument.