Allow spaces in domain lists passed to constraints
The current list of domains passed to the Domain Exclusion Constraint
has this format:

"host1, host2, host3"

This causes none of the domains being correctly parsed by the
constraint, as it does split the list based on the comma ",".
So the Domain Exclusion Constraint was translationg the list as:

["host1", " host2", " host3"]

Keeping the spaces between the commas and the first charater for the
next domain. This would cause requests to "host2" and "host3" not
matching the listed domains in the Constraint.

We could simply remove the spaces in the domain list in order to solve
this, but:
- We had no test coverage for this behaviour. Would be nice to
  explicitly cover this.
- I generally like to follow the Postel's Law:
  "be liberal in what you accept, and conservative in what you send"

So I decided to, instead of "fixing the list", allowing spaces between
the commans in the Domain constraints inputs, as the code changes to
achieve so were minimal and now we covered the situation with tests.