Default to Submit page when requesting localhost
Until now, when loading the Cosmetics application in our machines for
development, we needed to set up SUBMIT_HOST and SEARCH_HOST in our
environment variables, and access the Cosmetics in our machines by
using either of those values as URL.

Attempting to load the service by using: "http://localhost:300x" was
resulting into an error page.

For development purposes we want to default to the Submit page when
requesting the service from our computers with "http://localhost:300x".