Normalize assertion over notification tests
A couple tests became flaky when asserting over the flash message after
a notification has been submited.

I manually checked the flash message is pased to the controller and the
destination page correctly contains the submitted  notification, but we
do not have the flash message in the output page.

Decided to normalize the assertions between all the examples, asserting
over the actual listed notification title instead of asserting over the
flash message.