Default to Submit page when not specifically requesting Search pages (#1662)
* Default to Submit unless explicit Search requests

Instead of having:
- "Search routes"
- "Submit routes"
- "everything else"

We are opting for:
- "Search routes"
- "Everything else we will try to default it to Submit routes"

In This way the requested paths will be checked against Submit routes
unless they come under the specific Search domain/host.

* Normalize routes constraints naming

DomainCheck and DomainConstraint are basically the same domain check
but inverted:
- DomainCheck checks the exclusion of the host from a domain/list of
  domains.
- DomainConstraint checks the inclusion of the host ...

Given that, normalizing their names and turning them more explicit will
make clearer their function without having to check their internals.

* Cover Domain Exclusion Constraint with tests

The constraint class wasn't covered by unit tests.
The tests are basically the same as the ones written for the inclusion
constraint, but with the assertions expecting opposite results.

* Allow spaces in domain lists passed to constraints

The current list of domains passed to the Domain Exclusion Constraint
has this format:

"host1, host2, host3"

This causes none of the domains being correctly parsed by the
constraint, as it does split the list based on the comma ",".
So the Domain Exclusion Constraint was translationg the list as:

["host1", " host2", " host3"]

Keeping the spaces between the commas and the first character for the
next domain. This would cause requests to "host2" and "host3" not
matching the listed domains in the Constraint.

We could simply remove the spaces in the domain list in order to solve
this, but:
- We had no test coverage for this behaviour. Would be nice to
  explicitly cover this.
- I generally like to follow the Postel's Law:
  "be liberal in what you accept, and conservative in what you send"

So I decided to, instead of "fixing the list", allowing spaces between
the commans in the Domain constraints inputs, as the code changes to
achieve so were minimal and now we covered the situation with tests.

* Stub actual request host in spec domain helpers

We were stubbing the domain constraint checks methods without actual
correlation between the specific domain we wanted to simulate and the
actual stub.

This caused test brittleness once we modified how we use the domain
constraints for the routes.

Also, being used on request and feature tests, we should lean towards
end to end integration testing instead of method stubbing.

By just stubbing the request host, we allow end to end testing while
effectively faking which domain is being used during specific tests.

* Normalize assertion over notification tests

A couple tests became flaky when asserting over the flash message after
a notification has been submitted.

I manually checked the flash message is passed to the controller and the
destination page correctly contains the submitted  notification, but we
do not have the flash message in the output page.

Decided to normalize the assertions between all the examples, asserting
over the actual listed notification title instead of asserting over the
flash message.