Simplify JSON quotes trimming
We needed to pipe the output from "jq" into other commands. As "jq"
output was being wrapped between quotes, we had to trim those quotes
before using the string in subsequent commands.

"jq" offers a flag that generates the output without quotes. So we
can avoid having to pipe the output into a trim call.