Set GitHub deploys as inactive when pr merged (#1673)
* Simplify JSON quotes trimming

We needed to pipe the output from "jq" into other commands. As "jq"
output was being wrapped between quotes, we had to trim those quotes
before using the string in subsequent commands.

"jq" offers a flag that generates the output without quotes. So we
can avoid having to pipe the output into a trim call.

* Inactivate Review App GH deploys when closing PRs

On top of deleting the review app itself when a Pull Request is merged
or closed, we want to set the status of the corresponding Github
deployment as "inactive", so we won't have a "View Deployment" link for
an environment that does not exist anymore.