Fix request specs
Had to comment/ignore a couple specs as they check the "invalid-user"
case when we do requests on:
- search domain with a submit user.
- submit domain with a search user.

As currently we have Search users using Keycloak and Submit users
using Devise, we use different keycloak/Devise helpers depending
on the Search/Submit request domain. This does not allow to test
the flow for "visiting domain with wrong user".