Favour invalid account over Devise standard error
Since having moving to Devise, when trying to authenticate a search
service user on Submit url (or the other way around), was redirecting
users to the sign-in page as the authentication fails.

We want to keep the original behaviour instead, where the user gets
shown an "Invalid account" information page.

To solve that, we moved the controller before action hook checking for
invalid user types for the service to happen before we attempt to
authenticate the user.
This way the wrong users will be shown the "Invalid account" page
instead of being redirected to sign-in.