Fix user registration "email already taken"
When having validation errors on the registration form and submitting
the form for second time after correcting them, the service is throwing
an "email is already taken" error.

This happens because we placed the form validation in the a block call
for Devise #create method. The block gets executed after the resource
is saved, so the user is effectively created even if the form contains
issues on validation.
Seems this block is intented for "after creating user" actions.

Instead of using the method block, we are going to guard against
validation errors as a first step, and only call Devise#create if the
form is valid.