Fix service error when should display Invalid Link
For signed in users following an already used "reset password" link
for second time, the "you're already signed in as another user"
check was failing as no user with the given reset password token
was found. The reset password token was removed from the user once
the password was successfully reset the first time the link was
used.