Add workflow for deleting dangling review apps
Every night at 1am the action will run to identify and delete review
apps that are deployed but do not correspond with any active Pull
Request.

The review apps should be removed when their matching pull request is
closed, as we have another Github worflow for achieving so. This script
act as a regular cleaning for when that one/off action fails.