Fix resp person controller filter without 2FA
When having disabled the 2FA/Secondary Authentication option, the
filter redirecting to the create or join an existing Responsible
Person wasn't working properly and a newly registered user would be
left in a "broken" state.

To fix it, we want to ensure the mobile number of the user to be
verified before redirecting to create/join resp person ONLY when 2FA is
enabled, as the sms code from 2FA authentication is how an user gets its
mobile number verified.