Allow resp. pers. invitations from different users
Now that we record the user sending the responsible person invitation,
we will allow having multiple invitations to the same user for a single
team as long as each invitation comes from a different team member.

This way, we will be able to display the list of users who sent the
invitations when showing the pending responsible person invitations
list.