Show list of inviting users on expired RP invitation
When the list of pending Responsible Person invitations is displayed,
when the invitation (or all invitations for same RP) is expired, the
listed invitation will display a message informing of the invitation
expiration, followed by the names of all the users who sent the
invitations for that responsible person.

This way we will allow the invited user to contact any of them in
order to resend the invitation.