Extract complex logic from view into a presenter
The presenter becomes a bridge between the model data and the view.

We're not decorating a model object for the view, but doing full
transformations over a collection of data.
This is the reason why I've chosen a plain old ruby object to implement
the presenter instead of relying into the decorator pattern, that does
not fit so well in this case.

Also opted for not introducing Gems like Draper but to just write a
PORO.