Show inviting user in team members pending invite
When listing the team members of a responsible person, the name of the
user who created the invitation will be displayed when listing members
with a pending invitation.

This avoids invitations to the same email sent by different team members
being shown as duplicated in the list, as now will be obvious that two
different people invited the same email and either invitation can be
refreshed independently.

We may want to change this to display them as a single line and maybe
the Resend Invitation link to be associated with the most recent one,
but this is out of scope of this ticket and still needs to be decided
by the team.