Update inviting user name when re-sending invitate
When an user re-sends an invitation from the team members page, the
invitation will now belong to the user who clicked the re-send link.

This will affect the value in the DB, the information in the inviting
email and the information shown in the team members element
associated to the invitation.