Fix: Flag security completed when registering through invitation
When an user creates an account from an invitation it does fill its
name, mobile number and password.

With this data, the user have effectively completed their account
security, even when haven't done it from the account security page
(that is reached when not registered through an invitation link).

This was missing, and caused the declaration page not being displayed
for users who registered through invitation links.