Allow duplicating RP name when having expired invite
We only want to block the duplication of the name when the user has
active invitations to another responsible person with the same name.

If the invitation is expired and the user cannot get it resent, they
could get blocked without being able to neither join the team or
create a team by themselves.