Allow duplicated RP names (#1850)
* Allow duplicated RP names with some exceptions

Currently a user will get an error when they try to create an RP with
the same name (see attachment). This will prevent RPs from adding
their legal RP name that appears on any packaging.

Users need to be able to add their RP name as it appears on any
packaging (this is a legal requirement).
This means 3rd parties could prevent RPs from complying with the
legislation (e.g. Boots find that Boots has already been created and
then they can’t notify their products).

The error message should be removed and users be allowed to proceed to
creation of an RP, even if that same RP name already exists in the
database.

EXCEPTIONS.

We won't allow creating a RP with the same name as an existing one
when:
1. The user creating the RP already belongs to another RP with the
   same name.
2. The user creating the RP has been invited to another RP with the
   same name.