Setup CSP Header in Report Only mode
One of the outputs of a Pentest auditory over the service was the need
of setting the Content Security Police header to decrease the chances
of XSS attacks on the service users.

As starting point, we're enabling the CSP with a default configuration
and in "report only" mode. So the violations of the CSP will be reported
to Sentry, but they won't be blocked.