Add "nonce" to inline JS scripts on views
Having inline JS scripts in the views is discouraged as could be an open
door for XSS attacks.
By adding the "nonce" tag to the scripts, CSP will only allow inline
scripts that are tagged with the nonce (nonce gets re-generated per each
request), mitigating the risk of execution of injected XSS in the
codebase.