Setup Content Security Policy header for reporting (will not block resources) (#1852)
* Setup CSP Header in Report Only mode

One of the outputs of a Pentest auditory over the service was the need
of setting the Content Security Policy header to decrease the chances
of XSS attacks on the service users.

As starting point, we're enabling the CSP with a default configuration
and in "report only" mode. So the violations of the CSP will be reported
to Sentry, but they won't be blocked.

* Allow unsafe inline style src in CSP config

Not very problematic security-wise and would be a pain to identify
and extract.

* Add "nonce" to inline JS scripts on views

Having inline JS scripts in the views is discouraged as could be an open
door for XSS attacks.
By adding the "nonce" tag to the scripts, CSP will only allow inline
scripts that are tagged with the nonce (nonce gets re-generated per each
request), mitigating the risk of execution of injected XSS in the
codebase.