Bugfix: Files not being flagged from antivirus
Cause:

The analyzer metadata generates a hash with symbols as keys, while the
check against the metadata output was using strings as keys.

Because of the non existent key being queried, the check on metadata
safety was always returning null value, not false. Hence the
Notification File was never getting flagged with the "file flagged as
virus" upload error, even when the metadata indicated so.

Added both unit specs for the AV analyzer and a feature spec to cover
the virus detection flow end to end.

(cherry picked from commit 5fa4ddc89190a794121929d2b0ba083f770b5ee5)