Fix assumption that files with virus get deleted
The codebase around images uploads seemed to imply that when there is
no attachment in the image upload is because it failed the antivirus
check and the attachment got deleted.

After inspecting the behaviour of the service/antivirus, we noticed
that this assumption is wrong.
The antivirus check simply flags the safety failure on the file metadata.
It DOES NOT delete the attachment.

Based on that, we arem odigying the service logic to check against the
file metadata for the antivirus output instead of relying on the file
attachment existence.