Bugfix: Show add. info page while waiting for AV
When an user adds an image to the notification and gets redirected to
the additional information page, it gets sent back to the adding an
image page because the image is not "safe" by antivirus check.

The antivirus check runs asynchronously. So this will always happen
as the user submits the image and gets redirected to the Additional
Info page while the Antivirus check is scheduled in the background.

This generates an endless loop that never allows the user to actually
reach the additional info page.

In order to fix this behaviour, we are allowing to load the Additional
Information page when the image Antivirus check is pending.
User will see the pending state next to the image information and will
be able to refresh the page (to see the check output).