Bugfix: Files not being flagged from antivirus (#1871)
While working on refactoring bits of the antivirus-related code in other ticket, noticed that:

- We don't have specs covering neither the antivirus analyzer nor the end to end flow when detecting a virus in an user uploaded file.
- The tagging of files with virus error was not happening even when virus are detected. So the end-to-end flow of displaying a notification as "Error" when a virus is detected during the submission was broken.

Cause:

The analyzer metadata generates a hash with symbols as keys, while the
check against the metadata output was using strings as keys.

Because of the non existent key being queried, the check on metadata
safety was always returning null value, not false. Hence the
Notification File was never getting flagged with the "file flagged as
virus" upload error, even when the metadata indicated so.

Added both unit specs for the AV analyzer and a feature spec to cover
the virus detection flow end to end.