Ensure deleting Notification Files purges attachs.
When calling "delete" over an object that has an attachment, we're
deleting the object without triggering the callbacks associated to
its removal.
This leaves the attachment/blob files associated to the object
undeleted, becoming orphans as they're not associated to any
Notification File.

In order to trigger the complete cleanup of the Notification File and
its associated attachment/blob, we need to call "destroy" instead.