Ensure deleting Component purges its form. file
When calling "delete" over an object that has an attachment, we're
deleting the object without triggering the callbacks associated to
its removal.
This leaves the attachment/blob formulation file associated to the
object undeleted, becoming orphans as they're not associated to any
component.

In order to trigger the complete cleanup of the Component and its
associated formulation file, we need to call "destroy" instead.