Fix image upload page for flow with ZIP file containing a multi-component notification (#1872)
* Refactor controller by extracting methods

- Use of guard clause to reduce the nesting condition levels.
- Extract small methods with descriptive names to make the controller
  easier to follow for the readers.

* Refactor: Move logic from controllers to model

The logic for adding an image to a notification had some processing
that didn't belong to a controller class.
As we previously encapsulated the logic into a method, now we moved
the method to the proper object that should receive that message:
"Add image to the notification" --> Notification.add_image

* Refactor: ImageUpload virus check related methods

- Names adding human readable abstraction over metadata "safe" output
  from antivirus check.
- Enforced the pass/fail/pending possible states of the antivirus check
  output.
- Added Unit Tests.

* Fix factory returning wrong type

The antivirus scanner sets metadata to either true/false boolean types,
not a string.

* Fix assumption that files with virus get deleted

The codebase around images uploads seemed to imply that when there is
no attachment in the image upload is because it failed the antivirus
check and the attachment got deleted.

After inspecting the behaviour of the service/antivirus, we noticed
that this assumption is wrong.
The antivirus check simply flags the safety failure on the file metadata.
It DOES NOT delete the attachment.

Based on that, we arem odigying the service logic to check against the
file metadata for the antivirus output instead of relying on the file
attachment existence.

* Bugfix: Show add. info page while waiting for AV

When an user adds an image to the notification and gets redirected to
the additional information page, it gets sent back to the adding an
image page because the image is not "safe" by antivirus check.

The antivirus check runs asynchronously. So this will always happen
as the user submits the image and gets redirected to the Additional
Info page while the Antivirus check is scheduled in the background.

This generates an endless loop that never allows the user to actually
reach the additional info page.

In order to fix this behaviour, we are allowing to load the Additional
Information page when the image Antivirus check is pending.
User will see the pending state next to the image information and will
be able to refresh the page (to see the check output).

* Show multiple images upload page for multicomp.

When the Notification zip file ocontains a multicomponent Notification,
when reaching the "upload image" page, we will show the upload multiple
images page instead of the text for a single image.

* Fix display of multiple label images in overview

So the Refresh button is not in the same line as the next listed image.