Don't process Notification file unles virus free
The antivirus job and the notification file processing are both run
in async jobs in parallel.

This allows the file to be processed while is still being analyzed for
virus or, even worse, when it actually contains a virus file in the zip.

In order to avoid that we guard the process execution behind a check on
the antivirus output:
- If the antivirus process finished and the file is marked as safe,
  process the notification as usual.
- If the antivirus process hasn't run on the file, reschedule the
  processing of the notification to be run again later.
- If the antivirus process finished and detects a virus on the
  attachment, do not process the file.