Fix incomplete accounts trying to reset password
When an user with an incomplete account (missing their mobile number,
name or password) attempts a password recovery the service is throwing
an error.

Sadly seems some users get confused after introducing their
email in the first step of the Submit user registration and, instead
of checking their email for the confirmation link (as prompted), they
directly attempt to login, and not having set their password they
attempt to reset it.

The cause of the exception is the 2FA attempt prior to the password
reset, while the user actually doesn't have a mobile number to send the
2FA code to.

As this bit of the code is shared between Submit and Search users, we
are changing this behaviour to:

- SUBMIT USERS:
  When a submit user that didn't complete its registration attempts to
  reset its password (that is not set up), will be resent the account
  confirmation instructions instead.

- SEARCH USERS:
  When a search user that didn't complete its registration after having
  received an invitation link attempts to reset its password, will be
  resent the invitation link instead.