Avoid duplicated emails with different capitalisation
Users were able to register a new account with an email already
belonging to another user but with different capitalisation.

This was happening because when saving the user in the new account
form we were bypassing the validations for creating a partially
completed user. While the business logic to check if the user already
exists is using a case sensitive call (find_by) and not detecting
that there is already an user with that email.

This caused the user becoming invalid in a later stage when trying to
submit their security information to complete the registration.
At that point the email does not pass the validation and the form
submission triggers an unexpected error.