Restrict Responsible Person postcode to UK
It is a requirement that the Responsible Person is stablished in the UK.

Decided to implement this with a Regular Expression as should be good
enough to check UK Postcode format.
Could have gone for gems or consulting third party services to check for
real addresses but seems an overkill for our goal (don't allow random
or foreign values there).