Fix back link from new RP type selection
When an user that was belonging to multiple RP reached this page before
having selected their current RP, the back link route was throwing an
exception.

Fixed by pointing the back link to the RP selection when the user has
multiple RPs but none of them defined as their current one.