Remove Striping gem in favour of local Concern
Insteado of relying on an external gem for striping whitespaces from
model attributes, we removed the dependency and brought the behaviour
into an Rails Model Concern.

This works for both ApplicationRecord classes or classes using
ActiveModel without relying on DB table (Like rails form objects).
the list of attributes to be striped can be specified or we can rely on
the concern default behaviour:
- For ApplicationRecord classes strip attributes that have changed
  their value.
- For ActiveModel classes (not AR) strip all attributes.