Introduce stricter email format validation
We had multiple occurrences of clearly wrong emails being accepted by
our validations and failing when attempting to be used by Notify
service.

The validate_email gem was severely deprecated, and its validation was
so relaxed that was accepting emails like "user@example",
"user@example com", "user @ example . com"...
The only value of the gem was to define a custom email validator that
checked that the email address could be parsed and that the email
address domain is not local (user@example).

The approach taken was to remove the gem, create our own custom email
validator reproducing the gem behaviour and to extend the validator to
check the email against accepted email format REGEXP on top of the
domain check.

To validate the email format we are using URI::MailTo::EMAIL_REGEXP,
that is standard in Ruby library and strict enough to catch the majority
of typos user may be introducing.
The downside with this regexp is that accepts local domains, as they
technically are valid email addresses.

But, as we are combining both the REGEXP and the domain check in our
custom validator, we have both fronts covered.

The Ruby regexp has beel also set as the email format regexp used by
Devise. As Devise default one is waaaay too relaxed and allows multiple
typos.